finished v1

+++ file_handling.py
@@ -3,20 +3,19 @@ import sys
 class Argument(object):
 
     def __init__(self):
-        pass
+        self.num_lines = sum(1 for line in open("todo_list.txt"))
 
     def arg_without_arguments(self):
         print("\nCommand Line Todo application\n=============================\n\nCommand Line arguments:\n-l Lists all the tasks\n-a Adds a new task\n-r Removes a task\n-c Completes a task\n")
 
     def list_all(self):
-        num_lines = sum(1 for line in open("todo_list.txt"))
-        if num_lines == 0:
+        if self.num_lines == 0:
             print("\nYou don't have any tasks for today! :)\n")
         else:
             print("\nYour tasks are:\n")
             todo_list = open("todo_list.txt", "r")
             line_count = 1
-            for _ in range(num_lines):
+            for _ in range(self.num_lines):
                 print(str(line_count) + " - " + str(todo_list.readline()))
                 line_count += 1
             todo_list.close()
@@ -31,19 +30,18 @@ class Argument(object):
             print("Unable to add: no task provided.")
 
     def remove_task(self, item):
-        num_lines = sum(1 for line in open("todo_list.txt"))
         if len(item) == 0:
             print("Unable to remove: no index provided.")
-        elif type(int(item)) != int:
+        elif not item.isnumeric():
             print("Unable to remove: index is not a number.")
-        elif int(item) > num_lines:
+        elif int(item) > self.num_lines:
             print("Unable to remove: index is out of bound.")
         elif len(item) > 0:
             todo_list = open("todo_list.txt", "r+")
             lines = todo_list.readlines()
             item = int(item)
             todo_list.seek(0)
-            for i in range(num_lines):
+            for i in range(self.num_lines):
                 if i + 1 != item:
                     todo_list.write(lines[i])
             todo_list.truncate()
@@ -51,16 +49,23 @@ class Argument(object):
             print("Task removed successfully!")
     
     def complete_task(self, item):
-        num_lines = sum(1 for line in open("todo_list.txt"))
-        todo_list = open("todo_list.txt", "r+")
-        lines = todo_list.readlines()
-        item = int(item)
-        todo_list.seek(0)
-        for i in range(num_lines):
-            if i + 1 != item:
-                todo_list.write(lines[i])
-            else:
-                todo_list.write("[x]" + lines[i][3:])
-        todo_list.truncate()
-        todo_list.close()
-        print("Task checked successfully!")
+        if len(item) == 0:
+            print("Unable to check: no index provided.")
+        elif not item.isnumeric():
+            print("Unable to check: index is not a number.")
+        elif int(item) > self.num_lines:
+            print("Unable to check: index is out of bound.")
+        else:
+            num_lines = sum(1 for line in open("todo_list.txt"))
+            todo_list = open("todo_list.txt", "r+")
+            lines = todo_list.readlines()
+            item = int(item)
+            todo_list.seek(0)
+            for i in range(num_lines):
+                if i + 1 != item:
+                    todo_list.write(lines[i])
+                else:
+                    todo_list.write("[x]" + lines[i][3:])
+            todo_list.truncate()
+            todo_list.close()
+            print("Task checked successfully!")
